# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "轻记 · 本地概念默写",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/cards/{card_id}/history": {
      "get": {
        "tags": [
          "统计"
        ],
        "summary": "Card History",
        "operationId": "card_history_api_cards__card_id__history_get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "tags": [
          "统计"
        ],
        "summary": "Statistics",
        "operationId": "statistics_api_stats_get",
        "parameters": [
          {
            "name": "folder_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder Id"
            }
          },
          {
            "name": "tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "card_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer"
          },
          "time_spent_ms": {
            "type": "integer",
            "title": "Time Spent Ms"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "graded",
              "failed"
            ],
            "title": "Status"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "completeness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completeness"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "merged_feedback": {
            "additionalProperties": true,
            "type": "object",
            "title": "Merged Feedback"
          },
          "model_knowledge_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Model Knowledge Notes"
          },
          "judges": {
            "items": {
              "$ref": "#/components/schemas/JudgeResponse"
            },
            "type": "array",
            "title": "Judges"
          },
          "disagreement": {
            "type": "boolean",
            "title": "Disagreement"
          },
          "merge_fallback": {
            "type": "boolean",
            "title": "Merge Fallback"
          },
          "grading_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grading Error"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "error_types": {
            "items": {
              "$ref": "#/components/schemas/NamedType"
            },
            "type": "array",
            "title": "Error Types"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_id",
          "position",
          "term",
          "user_answer",
          "time_spent_ms",
          "submitted_at",
          "status",
          "definition",
          "accuracy",
          "completeness",
          "final_score",
          "merged_feedback",
          "model_knowledge_notes",
          "judges",
          "disagreement",
          "merge_fallback",
          "grading_error",
          "prompt_version",
          "model_name",
          "created_at",
          "error_types"
        ],
        "title": "AnswerResponse"
      },
...
      "CardHistory": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/CardOutput"
          },
          "stats": {
            "$ref": "#/components/schemas/CardStats"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerResponse"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "card",
          "stats",
          "answers"
        ],
        "title": "CardHistory"
      },
...
      "CardOutput": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "reference_note": {
            "type": "string",
            "title": "Reference Note"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/NamedOutput"
            },
            "type": "array",
            "title": "Tags"
          },
          "folders": {
            "items": {
              "$ref": "#/components/schemas/NamedOutput"
            },
            "type": "array",
            "title": "Folders"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "term",
          "definition",
          "reference_note",
          "tags",
          "folders",
          "created_at",
          "updated_at"
        ],
        "title": "CardOutput"
      },
      "CardStats": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/NamedOutput"
            },
            "type": "array",
            "title": "Tags"
          },
          "folder_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Folder Ids"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "average_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score"
          },
          "lowest_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lowest Score"
          },
          "last_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Score"
          },
          "average_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Time Ms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "term",
          "tags",
          "folder_ids",
          "attempt_count",
          "average_score",
          "lowest_score",
          "last_score",
          "average_time_ms"
        ],
        "title": "CardStats"
      },
...
      "ErrorDistribution": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "count"
        ],
        "title": "ErrorDistribution"
      },
...
      "ExamResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_name": {
            "type": "string",
            "title": "Folder Name"
          },
          "parent_session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Session Id"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "submitted_count": {
            "type": "integer",
            "title": "Submitted Count"
          },
          "graded_count": {
            "type": "integer",
            "title": "Graded Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "pass_threshold": {
            "type": "number",
            "title": "Pass Threshold"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "folder_id",
          "folder_name",
          "parent_session_id",
          "started_at",
          "finished_at",
          "total",
          "submitted_count",
          "graded_count",
          "failed_count",
          "pass_threshold",
          "answers"
        ],
        "title": "ExamResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JudgeResponse": {
        "properties": {
          "judge_index": {
            "type": "integer",
            "title": "Judge Index"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "completeness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completeness"
          },
          "raw_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Json"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "judge_index",
          "success",
          "accuracy",
          "completeness",
          "raw_json",
          "raw_text",
          "error"
        ],
        "title": "JudgeResponse"
      },
...
      "NamedOutput": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "NamedOutput"
      },
      "NamedType": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "NamedType"
      },
      "Overview": {
        "properties": {
          "card_count": {
            "type": "integer",
            "title": "Card Count"
          },
          "folder_count": {
            "type": "integer",
            "title": "Folder Count"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "graded_count": {
            "type": "integer",
            "title": "Graded Count"
          },
          "average_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score"
          },
          "average_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Time Ms"
          }
        },
        "type": "object",
        "required": [
          "card_count",
          "folder_count",
          "attempt_count",
          "graded_count",
          "average_score",
          "average_time_ms"
        ],
        "title": "Overview"
      },
...
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "card_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer"
          },
          "time_spent_ms": {
            "type": "integer",
            "title": "Time Spent Ms"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "graded",
              "failed"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_id",
          "position",
          "term",
          "user_answer",
          "time_spent_ms",
          "submitted_at",
          "status"
        ],
        "title": "QuestionResponse"
      },
...
      "StatsResponse": {
        "properties": {
          "overview": {
            "$ref": "#/components/schemas/Overview"
          },
          "error_distribution": {
            "items": {
              "$ref": "#/components/schemas/ErrorDistribution"
            },
            "type": "array",
            "title": "Error Distribution"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/CardStats"
            },
            "type": "array",
            "title": "Cards"
          },
          "recent_exams": {
            "items": {
              "$ref": "#/components/schemas/ExamResponse"
            },
            "type": "array",
            "title": "Recent Exams"
          }
        },
        "type": "object",
        "required": [
          "overview",
          "error_distribution",
          "cards",
          "recent_exams"
        ],
        "title": "StatsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}